+++ posthog/models/activity_logging/signal_handlers.py
@@ -5,6 +5,7 @@ from datetime import timedelta
 from django.contrib.auth import get_user_model
 from django.contrib.auth.signals import user_logged_in, user_logged_out
 from django.core.signing import TimestampSigner
+from django.db.models.signals import pre_delete
 from django.dispatch import receiver
 from django.http import HttpRequest
 
@@ -24,6 +25,7 @@ from posthog.models.activity_logging.activity_log import (
     changes_between,
     log_activity,
 )
+from posthog.models.experiment import ExperimentHoldout, ExperimentSavedMetric
 from posthog.models.organization_domain import OrganizationDomain
 from posthog.models.signals import model_activity_signal, mutable_receiver
 from posthog.models.user import User
@@ -293,4 +295,75 @@ def handle_organization_domain_change(
             name=detail_name,
             context=context,
         ),
+    )
+
+
+@mutable_receiver(model_activity_signal, sender=ExperimentSavedMetric)
+def handle_experiment_saved_metric_change(
+    sender, scope, before_update, after_update, activity, user, was_impersonated=False, **kwargs
+):
+    log_activity(
+        organization_id=after_update.team.organization_id,
+        team_id=after_update.team_id,
+        user=user,
+        was_impersonated=was_impersonated,
+        item_id=after_update.id,
+        scope="Experiment",  # log under Experiment scope so it appears in experiment activity log
+        activity=activity,
+        detail=Detail(
+            # need to use ExperimentSavedMetric here for field exclusions..
+            changes=changes_between("ExperimentSavedMetric", previous=before_update, current=after_update),
+            name=after_update.name,
+            type="shared_metric",
+        ),
+    )
+
+
+@receiver(pre_delete, sender=ExperimentSavedMetric)
+def handle_experiment_saved_metric_delete(sender, instance, **kwargs):
+    log_activity(
+        organization_id=instance.team.organization_id,
+        team_id=instance.team_id,
+        user=getattr(instance, "last_modified_by", instance.created_by),
+        was_impersonated=False,
+        item_id=instance.id,
+        scope="Experiment",  # log under Experiment scope so it appears in experiment activity log
+        activity="deleted",
+        detail=Detail(name=instance.name, type="shared_metric"),
+    )
+
+
+@mutable_receiver(model_activity_signal, sender=ExperimentHoldout)
+def handle_experiment_holdout_change(
+    sender, scope, before_update, after_update, activity, user=None, was_impersonated=False, **kwargs
+):
+    log_activity(
+        organization_id=after_update.team.organization_id,
+        team_id=after_update.team_id,
+        user=user or after_update.created_by,
+        was_impersonated=was_impersonated,
+        item_id=after_update.id,
+        scope="Experiment",  # log under Experiment scope so it appears in experiment activity log
+        activity=activity,
+        detail=Detail(
+            changes=changes_between(scope, previous=before_update, current=after_update),
+            name=after_update.name,
+            type="holdout",
+        ),
+    )
+
+
+@receiver(pre_delete, sender=ExperimentHoldout)
+def handle_experiment_holdout_delete(sender, instance, **kwargs):
+    from posthog.models.activity_logging.utils import activity_storage
+
+    log_activity(
+        organization_id=instance.team.organization_id,
+        team_id=instance.team_id,
+        user=activity_storage.get_user() or getattr(instance, "last_modified_by", instance.created_by),
+        was_impersonated=activity_storage.get_was_impersonated(),
+        item_id=instance.id,
+        scope="Experiment",
+        activity="deleted",
+        detail=Detail(name=instance.name, type="holdout"),
     )
